# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Plan-B API",
    "version": "0.2.0",
    "description": "Grasscutter Web Management Platform"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/quests": {
      "get": {
        "tags": [
          "Quests"
        ],
        "summary": "Script list + stats",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Scripts"
          }
        }
      }
    },
    "/api/quests/init-samples": {
      "post": {
        "tags": [
          "Quests"
        ],
        "summary": "Initialize sample scripts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Created count"
          }
        }
      }
    },
    "/api/quests/reload": {
      "post": {
        "tags": [
          "Quests"
        ],
        "summary": "Hot reload all scripts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}